+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 # CHANGELOG
 
+## Oct 3, 2020
+
+* Add support to retrieve datapoints
+
+## Oct 1, 2020
+
 * Initial timeseries GET